+++ tests/testthat/test-row_manipulate_df_rows.R
@@ -14,47 +14,47 @@ fgeo <- tribble(
 
 
 
-context("filter_status.R")
+context("row_filter_status.R")
 
 test_that("correctly handles wrong inputs", {
   not_chr <- 1
 
-  expect_error(filter_status("wrong"))
+  expect_error(row_filter_status("wrong"))
   expect_error(
-    filter_status(fgeo, "stem", not_chr)
+    row_filter_status(fgeo, "stem", not_chr)
   )
   expect_warning(
-    filter_status(fgeo, "stem", "wrong")
+    row_filter_status(fgeo, "stem", "wrong")
   )
   expect_error(
-    filter_status(fgeo, not_chr, "alive")
+    row_filter_status(fgeo, not_chr, "alive")
   )
   expect_error(
-    filter_status(fgeo, "wrong", "alive")
+    row_filter_status(fgeo, "wrong", "alive")
   )
   expect_error(
-    filter_status(fgeo, "stem", "alive", "not logical")
+    row_filter_status(fgeo, "stem", "alive", "not logical")
   )
 })
 
 test_that("returns data of correct class", {
-  out <- filter_status(fgeo, "stem", "alive")
+  out <- row_filter_status(fgeo, "stem", "alive")
   expect_true(any(grepl("data.frame", class(out))))
 })
 
 test_that("returns expected values", {
-  out <- filter_status(fgeo, "stem", "alive")
+  out <- row_filter_status(fgeo, "stem", "alive")
   expect_equal(
     unique(out$Status), "alive"
   )
 
-  out <- filter_status(fgeo, "stem", "dead")
+  out <- row_filter_status(fgeo, "stem", "dead")
   expect_equal(
     unique(out$Status), "dead"
   )
 
   w_status_tree <- fgeo %>% add_status_tree("alive", "dead")
-  out <- filter_status(w_status_tree, "tree", "dead")
+  out <- row_filter_status(w_status_tree, "tree", "dead")
   expect_equal(unique(out$status_tree), "dead")
   expect_equal(unique(out$Status), "dead")
   expect_equal(unique(out$CensusID), 1)
@@ -74,29 +74,29 @@ test_that("Warns when ignoring NA(s)", {
             2,    2,  "dead",
             2,    2,  NA
   )
-  expect_warning(filter_status(w_na, "stem", "dead"))
+  expect_warning(row_filter_status(w_na, "stem", "dead"))
 })
 
 
 
 
-context("stem_not_dead")
+context("row_keep_alive_stem")
 
-test_that("shortcut returns the same as filter_status", {
+test_that("shortcut returns the same as row_filter_status", {
   expect_equal(
-    filter_status(fgeo, "stem", "dead", exclude = TRUE),
-    stem_not_dead(fgeo, "dead")
+    row_filter_status(fgeo, "stem", "dead", exclude = TRUE),
+    row_keep_alive_stem(fgeo, "dead")
   )
 })
 
 
 
-context("tree_not_dead")
+context("row_keep_alive_tree")
 
-test_that("shortcut returns the same as filter_status", {
+test_that("shortcut returns the same as row_filter_status", {
   w_status_tree <- add_status_tree(fgeo, "alive", "dead")
-  long <- filter_status(w_status_tree, "tree", "dead", exclude = TRUE)
-  short <- tree_not_dead(w_status_tree, "dead")
+  long <- row_filter_status(w_status_tree, "tree", "dead", exclude = TRUE)
+  short <- row_keep_alive_tree(w_status_tree, "dead")
   expect_equal(long, short)
 })