+++ R/data.R
@@ -1,0 +1,23 @@
+#' Data sets for examples; adapted from [bciex::bci12vft_mini].
+#' @name datasets
+#' @examples
+#' library(dplyr)
+#' all_quad <- bciex::bci12vft_mini %>%
+#'   dplyr::rename(QX = x, QY = y) %>%
+#'   dplyr::filter(PlotID == 1) %>%
+#'   fgeo.tool::row_top(CensusID)
+#'
+#' top4quad <- all_quad %>% fgeo.tool::row_top(QuadratID, 4)
+#' identical(top4quad, fgeo.tool::top4quad)
+#'
+#' top1quad <- all_quad %>% fgeo.tool::row_top(QuadratID, 1)
+#' identical(top1quad, fgeo.tool::top4quad)
+#'
+#' top1quad
+NULL
+
+#' @rdname datasets
+"top4quad"
+
+#' @rdname datasets
+"top1quad"