+++ NEWS.md
@@ -6,6 +6,7 @@
     * `nms_*()`
     * `check_*()`
     * `ls_*()`
+    * `fieldforms_*()`
     
 * Add functions to deal with datasets with same variables but different names.
 * Add `str_suffix_match()`.